rename createJQuery to selectInContext, generalize type
The second (optional) argument to jQuery() specifies a selector
context (eg. DOM Element) to restrict the search within. By default
this is the doucment root.

This provides a more useful name for this use of the function, and
generalizes its type so that any Selectable can be used for either
argument.